refactoring map component
